requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.4
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
idna==2.8
lcs-client==1.3
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
